Represent money fields in Long cents, add delivery fee to order

+++ src/main/java/com/qelery/mealmojo/api/model/dto/MenuItemDto.java
@@ -8,7 +8,7 @@ public class MenuItemDto {
     private Long id;
     private String name;
     private String description;
-    private Double price;
+    private Long price; // cents
     private String imageUrl;
     private Boolean isAvailable;
 }
