+++ src/main/java/com/qelery/mealmojo/api/model/dto/RestaurantDtoIn.java
@@ -11,7 +11,7 @@ public class RestaurantDtoIn {
     private Integer pickupEtaMinutes;
     private Boolean deliveryAvailable;
     private Integer deliveryEtaMinutes;
-    private Double deliveryFee;
+    private Long deliveryFee; // cents
     private String logoImageUrl;
     private String heroImageUrl;
     private AddressDto address;
