+++ src/main/java/com/qelery/mealmojo/api/model/entity/Restaurant.java
@@ -25,7 +25,7 @@ public class Restaurant {
     private Integer pickupEtaMinutes;
     private Boolean deliveryAvailable;
     private Integer deliveryEtaMinutes;
-    private Double deliveryFee;
+    private Long deliveryFee; // cents
     private String logoImageUrl;
     private String heroImageUrl;
     private Boolean isActive = true;
